Finished conversion
It all works with our database now(hopefully). It didn't extensively
test the admin features, but they should all work because there weren't
changes for most of it.

When you open an empty cart it throws an error even though it is
correctly handled by an if statement. Is there a fix for that or should
we ignore it?