website+login
When we meet today we can smooth out our folders and delete the
unnecessary folders. This will be our final working folder and
everything inside here, website and login with the database works
correctly. From here on out all we need to the shopping cart integrated
and were free to make cosmetic changes